KanvasCamera: OpenGL camera preview
Step 1 on the path to Camera filters: render the camera in OpenGL!

This renders the camera preview on an OpenGL surface, by taking each
frame from the camera as a pixel buffer and using it as a texture, which
is applied to a surface through a fragment shader. This allows the
camera's preview to render with OpenGL, allowing us to change or layer
fragment shaders to do interesting thing like live filters, stickers
(even 3D!), and even using AR to use the camera with a 3D environment.
Cool stuff! Though, this change doesn't do any of that cool stuff; the
fragment shader used is just a passthrough, so there will be no visible
difference with this camera preview rendering.