KanvasCamera: OpenGL camera capture/recording
Now that the camera is being rendered with OpenGL, with the intention of
customizing the camera viewport outside of just the captured media, we
need to be able to capture the final rendering, rather than just the
camera output. This commit does exactly that for video and gifs. This
doesn't handle photos yet.