Kanvas Camera: Make filter selection carousel work
This commit adds icons to the filter selection carousel, makes them the
proper order, and makes it actaully apply a filter when selecting them.

NOTE: Chroma and Toon are just colors because we don't have an icon for
them yet.

Also rips out the previous left/right gestures and UILabel for applying
filters, and rips out the funky animation that was on the filters
selection carousel when the filter moves into the shutter. For the
animation, we'll implement it correctly in a later change.