KanvasCamera: fix saving filtered photos
Currently photos that are processed as pixel buffers fail to save as
JPGs, even though a UIImage exists and can be rendered in a UIImageView.
I suspect this is because UIImageView is robust in what it can render,
while UIImage.jpegData is not. Maybe to do with byte format mismatch
(RGBA vs. BGRA). Maybe it's because UIImageView computes the dimensions.
Either way, it's broken.

To fix it, I'm using `VTCreateCGImageFromCVPixelBuffer` from the
VideoToolbox module to do the conversion. Using CIImage (with a
CIContext + dimensions) also worked, but I perceived less memory usage
with VTCreate. This allows the image to be rendered and saved as a JPG.