KanvasCamera preview screen needs to chill the fuck out
When previewing captured media, the preview screen chews up a crazy
amount of memory. I first tried to fix it by stopping the camera session
when the preview screen opens, but that just made the AVPlayer stop.
Turns out the AVCaptureSession can't be stopped while an AVPlayer is
playing 🤔

Now that the capture session is stopped when viewing the preview, we
need to restart it if we back out of the preview screen.

Also, if you decide to save the media from the preview screen, currently
that dismisses, shows the camera for a few frames, and then dismisses
the camera. So that won't start the camera session again, just to close
it, we prevent the capture session from starting in this case.