KanvasCamera: fix crash and bad performance on iPhone 6
This commit fixes two issues:

1. A crash can happen when a frame is dropped with the OutOfBuffers
reason, because in that case we're resetting the filter view, but that
happens on another thread than the main thread, so if OpenGL tries to
use that filter after it's destroyed, it'll crash.

2. The iPhone 6 has horrible performance, since it seems to drop frames
with OutOfBuffers in a totally different way than all other devices. All
other devices seem to spew OutOfBuffers at a certain point, where the
iPhone 6 reports OutOfBuffers only a few times when the AVCaptureSession
is created. This means the existing recovery strategy of restarting the
AVCaptureSession actually makes things worse on the iPhone 6, causing
the camera to restart a bunch of times on startup, taking a long time to
load. Also, it's much easier to reproduce the aformentioned crash, and
other crashes.

To fix these two issues, I did a few things:

1. All access to filters are synchronized with the GLRenderer instance.

2. Recover from OutOfBuffers by tweaking the framerate, which causes the
AVCaptureSession to reallocate it's CMSampleBuffer cache.

3. If the main thread is taking too long to render a pixel buffer,
discard the buffer. This should help with keeping the camera output queue free
to continually provide frames.

4. Make sure GLRenderer callback queue (defaults to main queue) is being
used for all delegate calls.

Other changes that are unrelated, but... why not:

1. GLPixelBufferView was referencing `videoframe` uniform in Base.glsl,
but that doesn't exist. Rename it to the one that existed. This worked
before because of how textures are resolved as a uniform in a shader,
but it's not correct.

2. Call glDetachShader too.