Kanvas: store metadata on photos and videos created with camera
This stores metadata to indicate that the media was captured with the
Tumblr camera.

For photos, the EXIF `UserComment` field is used. For video, this
quicktime `Title` field is used (that's the only field that I could get
to work!). In both cases, just the string "Kanvas" is written. In the
future this will be a JSON document with info about how the media was
created and edited.

This also changes how media is moved to the photo library (initially
just in the example app) to ensure metadata comes with it.