Replace editor AVPlayer usage with GLPlayer
Instead of playing video with the AVPlayer, this adds an OpenGL-based
video player: GLPlayer. It reuses the camera's OpenGL rendering,
specifically GLPixelBufferView and GLRenderer, which enables the player
to render filters over the video. This also hooks up the editor's new
filter selector. This player also supports rendering still images, both
where the filter is static, or where it's dynamic and animates over the
image.

This just modifies the editor view; it doesn't actually save media with
filters from the editor yet. Next step ;)