Kanvas: fix flakey media picker test
A media picker test is flakey because the media picker button tries to
load the most recent image in the photo library as the button's
thumbnail, so it could be a different image, or not loaded yet. The
current snapshot test has the image as not loaded.

This adds the sample.png image to the library so the image is always the
same, and then waits two seconds to ensure it's loaded. The test then
deletes the image from the library.