Kanvas: fix camera loading
In the previous commit, when putting Kanvas in the swipable dashboard,
the camera wouldn't start properly, and would only function properly
when changing camera settings.

This is because the CameraController doesn't add any child view
controllers; it only adds their views. This means that the view
lifecycle methods don't get called properly, like
viewWillAppear/viewDidAppear/etc.

So, this commit primarily adds the CameraInputController as a child view
controller of CameraController. While there are other view controllers
that aren't proper children, that will be addressed in a subsequent change.

This commit also adds SplitViewController and TMPageViewController to
the KanvasCameraExample app, so that accessing Kanvas by swiping the
dashboard can be tested in the example app. It bypasses all the cool
stuff that was done in the example app to set feature switches, so I'll
want to fix that before cutting a PR.

Also, removes the Storyboard from the KanvasCameraExample app, because
fuck Storyboards.