Ensure status bar preferences are respected.
Currently, TMPageViewController and SplitViewController don't respect a
view controllers status bar hidden preference, so this commit fixes
that. However, I'm not hiding the status bar anywhere, since the
DashboardHeaderView doesn't respond properly to the status bar being
hidden after it's initialized. I'll fix that later, so for now the
status bar is visible everything, but I fixed some things so it's easier
to get working later.

Also, TMPageViewController in the KanvasCameraExample app was in the
same file as SplitViewController. Fix that too.