Kanvas: blur camera when transitioning views
This blurs the camera preview when the view is transitioning between
not focused and focused.

To do this, we can't just use UIKit hooks, we have to use
TMPageViewControllerDelegate to know exactly when swiping starts, and
when it's fully done. Also, to make sure that swiping is performant, we
probably shouldn't tear down the whole camera every time you hide it. To
accomplish this, we keep the camera off but still present for a set
amount of time, and then tear everything down afterwards. Right now it's
10 seconds, but that's probably too quick.

This adds this behavior to both Kanvas and KanvasCameraExample, which
is why there's some duplication (SplitViewController and
KanvasDashboardController). This commit also removes
TMPageViewController in KanvasCameraExample, since it's in SharedUI and
we already have a dependency on that module.