Kanvas: don't block the UI when starting and stopping the camera
Currently the camera session queue is only used with `sync` blocks. This
wasn't an issue with the camera previously, but now that the camera is
exposed from swiping the dashboard, the pause is very noticable.

The fix is to use `async` blocks, as there's no reason for using `sync`.
However, I've kept two cases using `sync`: one that recovers from out of
buffers, and another which is called from the NPF camera.