Kanvas: fix non-OpenGL camera.
An optimization for the OpenGL camera and the swipable dashboard ended
up breaking the non-OpenGL camera, since the AVCaptureSession isn't
created when the preview layer tries to use it. So, the best fix is to
just create the AVCaptureSession synchronously, but configure it
asynchronously (which is where the real work happens).