Kanvas fix clips from the media picker
Currently the media picker opening causes the CameraRecorder to be recreated,
which clears out the backing store for clips. This change moves the
backing store for clips (CameraSegmentHandler) to be owned by the
CameraController, and passed around to whatever needs it, so even if the
recorder is recreated for valid reasons, the clips will remain.