Kanvas: no async in deinit!
This commit fixes a EXC_BAD_ACCESS KERN_INVALID_ADDRESS in
kanvas.camera.sessionQueue when deallocating an object. This is because
the sessionQueue is being used in an async way, which shouldn't happen
from deinit.

Also, this adds the willCloseSoon check back to CameraInputController,
as that is needed to make sure the camera doesn't turn back on when
things are closing. This should improve stability on the iPhone 6.