Kanvas: resize & crop frames to match overlay
Currently the drawing overlay size doesn't match the media it's being
overlayed on, so it stretches to fit. This commit makes the editor
WYSIWYG, so we will actually crop and scale media to match the screen
size, and then overlay drawing on top of it.

This is only a temporary fix, as it makes video exporting very slow. A
better solution is to do the cropping and overlaying in OpenGL so it can
be entirely GPU-processed, rather than now it'll use the CPU to do
frame-by-frame operations.