Kanvas: refactor shaders to take a transformation matrix
All filters can now use a transform matrix to ensure media is properly
oriented. However, this doesn't work for media-picked videos just yet.

`GLPixelBufferView` also uses a transform matrix to make sure
media is rendered AspectFill (and this allows us to easily change that
in the future).

Also, it was weird that all the filters used `Filter.loadShader` to get
the shader's source code, and then passed that to `Shader.setProgram`.
Cleans this up with `Shader.getSourceCode`, and then the `Shader`
constructor takes the vertex and fragment source code.