Kanvas: fix seek assertionFailure
While Taichi was testing this PR, he saw this assertionFailure get hit,
since seeking probably fails _sometimes_ because the AVPlayerItem is not
readyToPlay. This doesn't happen on every device, but it seems like it
might be an iOS 12 vs 13 thing.

This assertionFailure doesn't provide any value in this case, so let's
just remove it.