premature optimization strikes again
When creating a filter chain to handle overlays, I was being clever and
trying to optimize the number of filters being applied (since each
filter allocates it's own rendering buffers). However, this breaks
overlays a bit, since the AlphaBlendFilter aplies it's transform in the
vertex shader, so it applies the transform to both the media and the
overlay, which isn't correct.

Now, always allocate a filter, even if it's off or passthrough, and then
allocate alpha blend filters for each overlay. This will only have an
effect when exporting video, as that's when the overlays are rendered in
OpenGL.