Kanvas: fix background fill (kinda)
Currently, background fill is VERY broken, as it only fills the area of
the image that was visible on screen, but since we render images as
AspectFill, some of the image is off-screen.

This change tracks when the user chooses to fill with a background
color, remembers that color, and draws that color outside the drawing
layer.

However, this isn't a perfect fix by any means. First off, it's a little
hacky. The ideal fix would be to make the drawing and sticker surfaces
as large as the media, and just position it appropriately. Filling would
just work then. Unfortunately, this is a large change that is too much
to finish in a few days. This commit will fix the bug "enough" for this
release. 806

"Wait, fix the bug `enough`"? LOL yes, well, this is broken with undo,
since we don't include the background fill in the undo history. That
being said, this is better than ALL usages of fill being broken.