Kanvas: prep to feature toggle OpenGL/Metal
The "OpenGL" directory is now called "Rendering", all the
OpenGL-specific classes are in a "OpenGL" sub-directory, and any class
that was prefixed with "GL" but doesn't necessarily depend on OpenGLES
has that prefix removed. Some non-OpenGL classes do use GLKit, which is
fine, since that usage is only for matrix structs like GLKMatrix4.

Dividing this up is in preparation for having Metal equivalents of the
OpenGL classes (really just Filter, Shader, and GLPixelBufferView).