Kanvas: tap and hold in GIF mode will double the recording time.
Currently in GIF mode, a 1 second video is recorded at 10fps, and then
looped to be a 2 second video. This commit allows for holding the
shutter to double the recording time to 2 seconds, and the looped the
final video will be 4 seconds.

This commit also ensures the shutter animation is the same duration as
the recording, and clears up some names.

This is a regression from 13.1 (!) when OpenGL rendering was added. It
wasn't a one-line fix because the shutter animation didn't exist then.