Kanvas: fix crash with emoji in editor
Whenn adding text to the editor, we figure out where line breaks,
erronous spaces, and blank lines are to draw the best looking
background. There was a bug with that implementation, where those lines
were being figured out with utf16 indexes, but used like they were
String offsets (which isn't utf16 encoded, they are emoji and grapheme
encoded as well). This caused a crash the minute an emoji was used,
since the index used to access the string would overflow.

This commit fixes this crash by always using the utf16 indexes against a
utf16 representation of the string, and then converting it back to a
string to account for emoji and grapheme properly.