Kanvas: fix Kanvas build and add camera permissions tests
Kanvas was broken because UIApplication.shared.open was used by a
module, and all modules are suppose to be extension-safe. So, move the
logic to the Kanvas app, and invoke it through delegation.

Also adds Camera permissions tests for the view (with snapshots) and the
controller. To test the controlled, the usage of AVCaptureDevice is
abstracted to CaptureDeviceAuthorizing