Kanvas: fix camera session start from NPF when first authorizing
Currently when opening the camera from the NPF media picker, the media
picker itself asks for camera permission. This is because the OLD camera
was part of the media picker. Now, even though the Kanvas camera is
seperate, we still prompt for access in the media picker. This should be
fine, but the problem is the media picker opens the camera immediate
after camera access is authorized, but then also asks for microphone
access. So, the camera opens (showing the permissions screen with camera
access granted, but mic access still needed), and then the microphone
access prompt shows, but when accepting the microphone access, the
permissions view has no idea it can change, so it doesn't update the
permissions screen. To make matters worse, the camera permissions
actively thinks this case is invalid, so when the user taps on the mic
access button, a developer assertionFailure happens, and then some
unexpected behavior.

This commit fixes that unexpected behavior to be correct (start the
camera and such). However, I've left the assertionFailure, since we
really don't want to be double-prompting for permissions. That'll be
fixed in another commit.