Kanvas: transform camera stream to screen aspect-ratio
While making the camera and editor aspect-fit, we lost the aesthetic of
having the camera being full-screen. To address this, this commit
transforms the camera stream to be the aspect-ratio of the screen. This
means the camera will render full-screen, as well as the editor (when
viewing media shot with the camera) but the editor will still render
other media in aspect-fit, so, best of both worlds!

To do this, we add a Filter to the beginning of the filter chain which
resizes and transforms the media to the proper aspect-ratio (and
orientation if it's a video with a rotation). Since each filter knows
its input and output dimensions, this means subsequent filters (effects
and overlays) get a properly sized and oriented input media.

There's one major issue with this approach that will be addressed in
subsequent commit: each filter allocates it's own pixel buffer pool.
This means it gets more memory intensive to add filters to the filter
chain. This can easily be fixed to add a PixelBufferProvider, which is
given to a Filter, to control the pixel buffer pool. Also, the size of
the pool should be changed based on the rendering usage. Ideally, we'd
only need to allocate one extra pixel buffer for each filter - to write
the final rendered media to. However, for live-rendering, we might need
a few extra in the buffer to not drop frames.