Kanvas: ungroup wave
Currently the "wave" filter is actually two different filters combined.
A filter has overhead, like allocating a pixel buffer to be the backing
store for the output texture, and needing to load, compile, and run
multiple shaders.

This commit makes the wave filter just one fragment shader so it can be
loaded by just one shader. This cuts it's memory utilization in 1/2.