Kanvas: fix color selector video glitch
The color selector pauses video to let you pan the screen and pick a
color. However, there's a bug in the Filter base class that doesn't
properly reset a stateful flag that controls the rotation of media, so
after the color is selected, dimensions of the media are out-of-sync
between the Renderer and Filter, and the media is rendererd improperly.

This commit fixes the bug by resetting the stateful flag:
switchedInputDimensions.