Kanvas: fix silent video export
Due to bundle paths being shifted around, one path went missing: the
path to silent.aac, which is used for making silent video (either with
photos from stitch mode, or frames from loops). This fixes the path, and
adds a KanvasCameraStrings.bundlePath(for:) method to keep this logic in
one place.