Resize GIF frames
Currently the GIFs that Kanvas exports are around 10-15MB (for a 2s
loop). This commit resizes GIF frames to try to keep the total GIF size
for a 2s loop around 1MB. I used hand-wavy math to come up with groups
of 400px and a fibonacci scale for scaling the image down. Also, the
scale depends on the screen scale factor, so the scale operates
properly.