Kanvas: load a GIF as image segments
Kanvas already supports playback of frames with a delay from the Stitch
camera. However, the Loop camera doesn't use this - it converts it
directly to a video. So, we have these two options for supporting GIFs,
and this commit implements the first one. We may also try going straight
to video if this is too memory intensive to have 50-ish GIF frames
loaded.

This uses CGImageSourceCreateImageAtIndex to get each frame in a GIF.
iOS 13 has a new API for GIF loading (CGAnimateImageAtURLWithBlock), but
since it only is present in iOS 13, I'm using
CGImageSourceCreateImageAtIndex for iOS 12 compatibility. Also, I've
seen color palette issues with CGAnimateImageAtURLWithBlock, so for now
this is the best in-the-box way of loading GIFs.