Kanvas: delay converting image to video
Currently, any time you add an image segment, the image gets converted
to a video and saved to disk, as we'll need the video when concatinating
the segments into a final video.

However, when importing a GIF (or any image with more than one frame),
it's quite expensive to convert each frame to a video and write it to
disk, especially if you end up not exporting it (the videos will be
deleted eventually, but still, it takes time).

This commit generates a video (during export) for any images that don't
have one. In practice, either all images will have videos (Stitch mode)
or no images will have videos (importing GIFs), but this can support any
combination of.

This commit also cleans up some of the media picking logic, and makes
sure to only try to load frames if there's more than one.