Kanvas: don't create a video clip by default for images
When capturing or picking images in Stitch mode, we generate a video
clip ahead of time as well, to use when making Stitch videos. This is
fine for capturing or picking single images, but when importing a GIF,
this can impact performance, as there can easily be 30+ frames. So, for
GIFs, we don't make these videos, and the process that will merge
segments will handle this for us. This already happens... so WTF does
this commit do?

This commit disables this ahead-of-time video creation everywhere, since
it just makes sense to always do this later.