Kanvas: fix GIF loading for some GIFs
The UIImagePickerController sometimes provides GIFs in the `.imageURL`
property when picking, but sometimes provides the first frame of the GIF
as a JPEG file. It's unclear why this happens, but looks like a bug with
UIImagePickerController.

To work around this, we use PhotoKit to request the original image, load
that image's data, and attempt to decode it as a GIF. If it works and it
has more than one frame, then we load it as a GIF.

This also cleans up URL generation and saving in Kanvas, as we had
variny logic on how to generate URLs. And everywhere used NSURL. Now
it's all URL, and all logic for how Kanvas makes URLs is in one place.