Kanvas: GIF Maker Preview
The new Kanvas GIF Maker now can import a GIF, or convert a video to a
GIF, and the tools will update the playback of the media in the editor.

The trimming tool will show thumbnails, but the timescale of thumbnails
to what actually gets trimmed is not correct - that will be finished in
another commit.