Kanvas: GIF Maker Export
This uses the settings applied in the GIF Maker (trim, speed, looping)
when exporting GIFs.

To do this, the MediaExporter gets a export(frames:...) overload, which
computes filters and overlays for each frame, and then
exports those frames as a GIF.

This only applies when every CameraSegment is an image, as the GIF Maker
converts everything to that. This is for backwards compatibility with
the GIF toggle button, as well as other types of multi-clip media
(Stitch mode), as that has it's own logic for delay between frames.