Kanvas: export GIFs to NPF
This commit implements didFinishExportingFrames in
PostFormsComposeFlowController, which will replace media in the edited image or
video block with the exported GIF, and save the GIF to the Photo Library.

This commit changes the method signature of didFinishExportingFrames to
include the dimensions of the GIF, and moves the size-computing logic to
the GIFDecoder.