Fix GIF export memory pressure
This commit fixes a serious memory pressure issue with exporting GIFs
from Kanvas.

Before, each frame of a GIF was being rendered by queuing it the default
qos global queue, and then being joined together when all tasks were
done. This was unnecessary, as there was no parallel tasks. This is also
quite memory intensive, as we allocate a closure for each frame, so
every frame is in memory.

This commit just iterates through each frame and exports it. This works
because exporting an image is a synchronous task. One task is queued on
the default qos global queue. This avoids needing to keep every frame in
memory.