Fix GIF Maker playback mode backgrounding
Backgrounding and foregrounding the app when having the GIF maker open
makes the playback mode selector view select the wrong option. This is
happening because auto layout and frame layout is being used together.
This commit fixes this by using constraints to position the selection
view, rather than it's frame.