Editor will default to GIF rebound mode when opening media captured from the GIF camera
Before the GIF Maker, the GIF camera was applying a rebound effect when
it was generating the output video. Now that we have the GIF Maker, we
can use it's rebound effect instead. This allows someone to turn it off
if they don't want, as before it was baked into the video.

While the GIF camera doesn't rebound, it still generates a video. A
subsequent change should be made to actually generate a GIF.