Make editor menu -> confirm button animation reliable
If the editor view doesn't animate the tools to the confirmation
checkbox, this commit now still calls the completion handler, but passes
a flag to indicate whether it animated or not. This is a practical
change because there's a valid case where the EditionMenuCollectionCell
provided isn't in the view hierarchy, so it has no parent view, and
cannot be animated. Rather that silently fail and not reaching the final
animation state, this allows it to just fail the animation, but still
"not be broken".