Fix time filters on GIFs
When exporting media, we were passing the frame delay as the effect
timestamp, which is wrong - that's just the delay between each frame,
which will usually be the same between each frame, and just make for a
static effect. This commit properly accumulates the time so the time
interval for effects is correct.