Toggle between Still and Live Photo as a GIF
When you select a Live Photo from the Create media picker, we grab both
the still image, and the paired video, but make a CameraSegment.image,
so it initially renders as a still image.

However, when the GIF button is pressed, that CameraSegment is changed
into a video, so the paired video then gets converted into a GIF. This
segment transformation is ephemeral, so when you revert the GIF, you go
right back to the still image.