Kanvas: fix GIF trimmer scrubbing
When dragging the trim handles, the GIF should scrub to show the frame
your moving the handle to. When dragging the entire GIF trim window, the
GIF should scrub to show the beginning of that window.

This commit fixes the end-handle scrubbing, as it wouldn't reliably show
the last frame. This also fixes the dragging-window scrubbing, as that
was totally pausing the playback of the GIF, due to the timer not being
added to the run loop.