Fix initial GIF trim settings
This is a little wee hack, but in the best interest of fixing this bug
for 16.7, I'm OK with it...

The bug is that the media player doesn't respect any initial GIF
trimming. So, if you import a long video, it won't trim it to three
seconds, it'll just convert the whole thing to a GIF.

However, from reading the code, and looking at tests for
GifMakerHandler, it seems like this bug shouldn't exist.

The bug is due to how the Gif Maker and the Media Player interact. The
Gif Maker loads frames, and then decides initial settings for trim,
speed, and playback. These settings are used to configure the Media
Player. But, then the Media Player is stopped and started again, which
is necessary, but, it causes the trim settings to be reset.

There's a few ways to fix this. While I'd ideally like to decouple
loading media from playing media, that's a larger change. So, this
commit just reconfigures the media player immediately after playing.
Since this happens on the main queue, there's no perceivable difference.

I call this a "little wee hack" since we're double-configuring the media
player to undo the resetting of settings that playback causes.