Fix crash when rebounding < 3 frames
The GIF Maker crashes when rebounding < 3 frames, because the Array
indexer is used for slicing with a range, in this case, we generate an
impossible range (1...0). Swift throws an exception in this case.

While this crash could be fixed by some ahead-of-time bounds checking,
we have a more elegant solution: dropFirst and dropLast. These methods
do what they say, but if there aren't enough elements, they just return
an empty array. So, in this case, we'll generate a rebound which is
equivalent to the original GIF. Which is the right thing to do, as a GIF
with two frames is also a rebound ;)