Ensure camera capture session is off when the camera is closed.
This PR ensures the camera capture session (AVCaptureSession) is stopped
whenever the camera is no longer in view. Or, more specifically, that
the capture session doesn't *start* when foregrounding... when the camera
isn't opened... o_0

This bug has existed since we introduced Create, but went
undetected until iOS 14's camera indicator. Now, if you open Create just
once, any time the app goes to the foreground we start the camera
capture session, even when the camera isn't visible (!) since the
CameraController is always loaded.

This commit fixes this by only registering for background/foreground
notifications when the view controller is visible, and unregistering
when it is no longer visible. Now the camera capture session won't turn
on without the view already being present.