Remove Differ pod and use Swift .difference(from:) instead.
The Differ pod was used in ManagedObjectCollectionDiffMaker to calculate changes in collections of NSManagedObjects.
This was updated to use Swift's native .difference(from:) method.
I removed the .move case from the CollectionChange enum and the moveRow from CollectionChangeSetApplier because in order to calculate the moves, we would need to call .inferringMoves() to the result of the calculated difference; this would result in extra cost and the delete and insert would be calculated anyway.
If we want to keep the moveRow though, we can use the .inferringMoves() function and make sure to call .move only once for delete or insert.