Remove assertions in OpenGL and replace with error logs
These assertions do not appear to actually guard against a critical issue. The WordPress iOS app is hitting them in debug builds but not in testing or production builds and exporting works properly. There may be issues with filtering or other effects, but these are not in use by all Kanvas users so shouldn't be blocked by an assertion here.